Bail out on a zero-sized buffer and cast err for %ld
snprintf writes nothing at all when bufsz is 0, not even a terminator, so the
checksum helper would run strchr over whatever the buffer already held. Return
before touching it.

int32_t is not long on every target, so cast before formatting with %ld.

+++ src/platform/stm32wl/STM32_LittleFS.cpp
@@ -273,7 +273,7 @@ const char *dbg_strerr_lfs(int32_t err)
 
     default:
         static char errcode[13];
-        snprintf(errcode, sizeof(errcode), "%ld", err);
+        snprintf(errcode, sizeof(errcode), "%ld", (long)err);
         return errcode;
     }
